pdf_render/pdf_render/requirements.txt
amqp==2.5.2
appnope==0.1.0
autopep8==1.4.4
backcall==0.1.0
billiard==3.6.1.0
boto==2.49.0
boto3==1.10.18
botocore==1.13.18
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2
django-cors-headers==3.7.0
django-utils-six==2.0
django-crontab==0.7.1
django-filter==2.4.0
django-modeltranslation==0.17.2
django-queryinspect==1.1.0
django-redis==4.10.0
django-rest-framework==0.1.0
django-s3-storage==0.12.5
django-ses==0.8.12
django-silk==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
drf-yasg==1.20.0
entrypoints==0.3
flake8==3.7.9
flex==6.14.1
future==0.18.2
gprof2dot==2017.9.19
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
jsondiff==1.2.0
jsonpickle==1.2
jsonpointer==2.0
kombu==4.6.6
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
openapi-codec==1.3.2
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow
pip-autoremove==0.9.1
pre-commit==1.20.0
prompt-toolkit==2.0.10
psycopg2-binary
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
raven==6.10.0
redis==3.3.11
requests==2.22.0
rest-condition==1.0.3
rfc3987==1.3.8
rollbar==0.14.7
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
simplegeneric==0.8.1
six==1.15.0
sqlparse==0.3.0
strict-rfc3339==0.7
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.7
validate-email==1.3
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0
pdf2image
poppler-utils
